+++ src/FiscalYears/EditFiscalYear/EditFiscalYear.js
@@ -9,18 +9,18 @@ import { stripesConnect } from '@folio/stripes/core';
 import { Paneset } from '@folio/stripes/components';
 import {
   LoadingPane,
-  useShowToast,
+  useShowCallout,
 } from '@folio/stripes-acq-components';
 
 import {
-  FISCAL_YEAR_VIEW_ROUTE,
-} from '../../../common/const';
+  FISCAL_YEAR_ROUTE,
+} from '../../common/const';
 import {
   fiscalYearResource,
-} from '../../../common/resources';
+} from '../../common/resources';
 import FiscalYearForm from '../FiscalYearForm';
 
-const EditFiscalYear = ({ resources, mutator, match, history }) => {
+const EditFiscalYear = ({ resources, mutator, match, history, location }) => {
   const fiscalYearId = match.params.id;
 
   useEffect(
@@ -32,13 +32,17 @@ const EditFiscalYear = ({ resources, mutator, match, history }) => {
     [fiscalYearId],
   );
 
-  const showToast = useShowToast();
+  const showCallout = useShowCallout();
 
   const closeEdit = useCallback(
     () => {
-      history.push(`${FISCAL_YEAR_VIEW_ROUTE}${fiscalYearId}?layer=view`);
+      history.push({
+        pathname: `${FISCAL_YEAR_ROUTE}/${fiscalYearId}/view`,
+        search: location.search,
+      });
     },
-    [fiscalYearId, history],
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [fiscalYearId, location.search],
   );
 
   const saveFiscalYear = useCallback(
@@ -46,7 +50,9 @@ const EditFiscalYear = ({ resources, mutator, match, history }) => {
       try {
         const savedFiscalYear = await mutator.fiscalYearEdit.PUT(fiscalYearValues);
 
-        showToast('ui-finance.fiscalYear.actions.save.success');
+        showCallout({
+          messageId: 'ui-finance.fiscalYear.actions.save.success',
+        });
         setTimeout(() => closeEdit(), 0);
 
         return savedFiscalYear;
@@ -60,7 +66,10 @@ const EditFiscalYear = ({ resources, mutator, match, history }) => {
         } catch (parsingException) {
           errorCode = 'genericError';
         }
-        showToast(`ui-finance.fiscalYear.actions.save.error.${errorCode}`, 'error');
+        showCallout({
+          messageId: `ui-finance.fiscalYear.actions.save.error.${errorCode}`,
+          type: 'error',
+        });
         throw new SubmissionError({
           _error: 'FY was not saved',
         });
@@ -101,6 +110,7 @@ EditFiscalYear.propTypes = {
   resources: PropTypes.object.isRequired,
   mutator: PropTypes.object.isRequired,
   match: ReactRouterPropTypes.match.isRequired,
+  location: ReactRouterPropTypes.location.isRequired,
   history: ReactRouterPropTypes.history.isRequired,
 };